+++ src/main/java/io/dataspaceconnector/model/auth/package-info.java
@@ -13,17 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.dataspaceconnector.service.usagecontrol;
-
-import org.springframework.stereotype.Component;
-
 /**
- * A {@link PolicyVerifier} implementation that simply allows access.
+ * Contains model classes for backend authentication.
  */
-@Component
-public final class AllowAccessVerifier implements PolicyVerifier<AccessVerificationInput> {
-    @Override
-    public VerificationResult verify(final AccessVerificationInput input) {
-        return VerificationResult.ALLOWED;
-    }
-}
+package io.dataspaceconnector.model.auth;