+++ src/test/java/io/dataspaceconnector/service/message/handler/ArtifactRequestHandlerTest.java
@@ -23,10 +23,12 @@ import de.fraunhofer.iais.eis.RejectionReason;
 import de.fraunhofer.iais.eis.TokenFormat;
 import de.fraunhofer.ids.messaging.handler.message.MessagePayloadInputstream;
 import de.fraunhofer.ids.messaging.response.ErrorResponse;
+import io.dataspaceconnector.camel.ClearingHouseLoggingProcessor;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.annotation.DirtiesContext;
 
 import javax.xml.datatype.DatatypeFactory;
@@ -40,6 +42,9 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 @SpringBootTest
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
 class ArtifactRequestHandlerTest {
+
+    @MockBean
+    ClearingHouseLoggingProcessor clearingHouseLoggingProcessor;
 
     @Autowired
     ArtifactRequestHandler handler;